Support attachment without ContentDisposition WIP
When a message/rfc822 is found whitout any Content-Disposition, the
parser was crashing. Now it no longer crash, but the attachment is not
recognize. This is a WIP. We need to adress the parsing of this type of
attachment. The attachment is a message/rfc822 which contains alternate
views. All the parts have the same 'id'.